Check for greyscale when small image uses palette
When a small file is stored with a palette, a test is made to see if
removing the palette, and the overhead that goes with it, results in
an even smaller file.  That test was always attempting RGB or RGBA
mode even when the image in question was greyscale.  I added a check
for greyscale.

(cherry picked from commit 537ca9bf78066f5267b5aec9813638ed1d1e8071)